机器人配置修改；统计访问量改显示当天的访问量，而不是至当天的访问量
git-svn-id: https://svn.chsi.com.cn/svn/repos/chsicommunity/knowledge@19512 5607bd42-195c-0410-82fb-b7440d194b30

+++ trunk/knowledge/src/main/java/com/chsi/knowledge/dao/CommonDAO.java
@@ -1,5 +1,6 @@
 package com.chsi.knowledge.dao;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.List;
 
@@ -34,5 +35,5 @@ public interface CommonDAO {
     List<SearchLogData> getSearchLogData(Calendar startTime, Calendar endTime);
     
     
-    List<Long> totalVisitLog(String kowId, String startTime, String endTime);
+    List<BigDecimal> totalVisitLog(String kowId, String startTime, String endTime);
 }
